Refactor: * Converters -> `converter` package * Rest files -> `utilities` package
