+++ src/main/java/edu/ucsd/sbrg/escher/gui/EscherConverterUI.java
@@ -53,9 +53,9 @@ import de.zbit.io.filefilter.SBFileFilter;
 import de.zbit.sbml.gui.SBMLWritingTask;
 import de.zbit.util.prefs.SBPreferences;
 import de.zbit.util.prefs.SBProperties;
-import edu.ucsd.sbrg.escher.EscherIOOptions;
+import edu.ucsd.sbrg.escher.utilities.EscherIOOptions;
 import edu.ucsd.sbrg.escher.models.EscherMap;
-import edu.ucsd.sbrg.escher.EscherOptions;
+import edu.ucsd.sbrg.escher.utilities.EscherOptions;
 
 /**
  * @author Andreas Dr&auml;ger
