+++ src/main/java/edu/ucsd/sbrg/escher/utilities/EscherOptions.java
@@ -12,7 +12,7 @@
  * <http://www.gnu.org/licenses/lgpl-3.0-standalone.html>.
  * ---------------------------------------------------------------------
  */
-package edu.ucsd.sbrg.escher;
+package edu.ucsd.sbrg.escher.utilities;
 
 import java.util.ResourceBundle;
