+++ src/main/java/edu/ucsd/sbrg/escher/utilities/EscherParser.java
@@ -14,7 +14,7 @@
  * <http://www.gnu.org/licenses/lgpl-3.0-standalone.html>.
  * ---------------------------------------------------------------------
  */
-package edu.ucsd.sbrg.escher;
+package edu.ucsd.sbrg.escher.utilities;
 
 import static org.sbml.jsbml.util.Pair.pairOf;
